+++ lib/electr/repl/printer.rb
@@ -3,13 +3,31 @@ module Electr
   # Prints the computation.
   class Printer
 
-    # result - Anything that respond to `truncate` and `round`.
-    #
-    # One can certainly benifit from a new type, say ElectrValue, as the
-    # type of `result`.
+    # result - The ElectrValue result of an evaluation.
     def self.run(result)
-      truncated = result.truncate
-      puts result == truncated ? truncated : result.round(10)
+      new(result).print
+    end
+
+    def initialize(result)
+      @result = result
+    end
+
+    def print
+      if @result.number?
+        print_number
+      else
+        print_error
+      end
+    end
+
+    def print_number
+      number = @result.number
+      truncated = number.truncate
+      puts number == truncated ? truncated : number.round(10)
+    end
+
+    def print_error
+      puts @result.error
     end
 
   end